+++ lab1/student.cpp
@@ -7,7 +7,7 @@ bool student::operator==(const student &other) const {
     return age_ == other.age_ && name_ == other.name_;
 }
 
-unsigned long long studentHasher::operator()(const student &student) {
+size_t studentHasher::operator()(const student &student) {
     // djb2 hash func
     std::string s = student.name_ + std::to_string(student.age_);
     unsigned long hash = 5381;
